refactor: code re-arrange
