+++ src/components/ProductCard/ProductCard.component.tsx
@@ -1,37 +1,7 @@
-import {
-	Avatar,
-	Badge,
-	createStyles,
-	Paper,
-	Text,
-	ThemeIcon
-} from '@mantine/core'
+import { Avatar, Badge, Paper, Text, ThemeIcon } from '@mantine/core'
 import { IconThumbUp } from '@tabler/icons'
 
-const ICON_SIZE = 20
-
-const useStyles = createStyles(theme => ({
-	card: {
-		position: 'relative',
-		overflow: 'visible',
-		display: 'flex',
-		flexDirection: 'column',
-		alignItems: 'center',
-		padding: theme.spacing.xl,
-		paddingTop: theme.spacing.xl * 1.5 + ICON_SIZE / 2
-	},
-
-	icon: {
-		position: 'absolute',
-		top: -ICON_SIZE / 3,
-		left: `calc(50% - ${ICON_SIZE / 2}px)`
-	},
-
-	title: {
-		fontFamily: `Greycliff CF, ${theme.fontFamily}`,
-		lineHeight: 1
-	}
-}))
+import { ICON_SIZE, useStyles } from './ProductCard.styles'
 
 interface IProductCard {
 	name: string
