+++ src/hooks/usePosts/usePosts.ts
@@ -1,40 +1,8 @@
 import { gql } from 'graphql-request'
 import { useInfiniteQuery } from 'react-query'
 
+import { IDataPosts, IProduct, ORDER } from '~/hooks/usePosts/types'
 import { graphQLClient } from '~/utils/graphQlClient'
-
-interface INode {
-	id: string
-	name: string
-	description: string
-	votesCount: number
-	thumbnail: {
-		url: string
-	}
-}
-
-type IDataPosts = {
-	posts: {
-		edges: Array<{ node: INode }>
-		pageInfo: {
-			endCursor: string
-			hasNextPage: boolean
-			hasPreviousPage: boolean
-			startCursor: string
-		}
-	}
-}
-
-export enum ORDER {
-	RANKING = 'RANKING',
-	NEWEST = 'NEWEST'
-}
-
-interface IProduct {
-	order: ORDER
-	page: number
-	after?: string
-}
 
 async function fetchProjects(variables: IProduct) {
 	const query = gql`
@@ -67,8 +35,8 @@ async function fetchProjects(variables: IProduct) {
 export function useProducts(order: ORDER, page: number, after = '') {
 	return useInfiniteQuery<IDataPosts, Error>({
 		queryKey: ['projects', order],
-		queryFn: data =>
-			fetchProjects({ order, page, after: data.pageParam || after }),
+		queryFn: ({ pageParam }) =>
+			fetchProjects({ order, page, after: pageParam || after }),
 		getPreviousPageParam: firstPage =>
 			firstPage.posts.pageInfo.hasPreviousPage
 				? firstPage.posts.pageInfo.startCursor
