+++ src/pages/projects/dark.tsx
@@ -1,29 +1,33 @@
 import { Fragment, useEffect, useState } from 'react'
 
-import { Grid, LoadingOverlay, Tabs } from '@mantine/core'
+import { Grid, Loader, Tabs } from '@mantine/core'
 import { IconFlame, IconNews } from '@tabler/icons'
 import { useInView } from 'react-intersection-observer'
 
 import { ProductCard } from '~/components/ProductCard/ProductCard.component'
-import { ORDER, useProducts } from '~/hooks'
+import { useProducts } from '~/hooks'
+import { ORDER } from '~/hooks/usePosts/types'
 
 export default function DarkTest() {
-	const [activeTab, setActiveTab] = useState<ORDER>(ORDER.RANKING)
-	const { data, isError, isLoading, fetchNextPage } = useProducts(activeTab, 20)
 	const { ref, inView } = useInView()
+	const [activeTab, setActiveTab] = useState<ORDER>(ORDER.RANKING)
+	const { data, isFetchingNextPage, isLoading, fetchNextPage } = useProducts(
+		activeTab,
+		20
+	)
 
 	useEffect(() => {
 		if (inView) {
 			fetchNextPage()
 		}
-	}, [inView])
+	}, [fetchNextPage, inView])
 
 	function generateProductCards() {
 		return (
 			<Grid>
 				{data?.pages.map(page => (
 					<Fragment key={page.posts.pageInfo.startCursor}>
-						{page.posts.edges.map(({ node }) => (
+						{page.posts.edges.map(({ node }, index) => (
 							<Grid.Col sm={12} md={3} key={node.id}>
 								<ProductCard
 									name={node.name}
@@ -40,28 +44,28 @@ export default function DarkTest() {
 	}
 
 	return (
-		<Tabs
-			value={activeTab}
-			onTabChange={setActiveTab}
-			style={{ position: 'relative' }}
-		>
-			<LoadingOverlay visible={isLoading} />
+		<Tabs value={activeTab} onTabChange={(tab: ORDER) => setActiveTab(tab)}>
 			<Tabs.List>
-				<Tabs.Tab value='RANKING' icon={<IconFlame size={14} />}>
+				<Tabs.Tab value={ORDER.RANKING} icon={<IconFlame size={14} />}>
 					Popular
 				</Tabs.Tab>
-				<Tabs.Tab value='NEWEST' icon={<IconNews size={14} />}>
+				<Tabs.Tab value={ORDER.NEWEST} icon={<IconNews size={14} />}>
 					Newest
 				</Tabs.Tab>
 			</Tabs.List>
 
-			<Tabs.Panel value='RANKING' pt='xl'>
+			<Tabs.Panel value={ORDER.RANKING} pt='xl'>
 				{generateProductCards()}
 			</Tabs.Panel>
 
-			<Tabs.Panel value='NEWEST' pt='xl'>
+			<Tabs.Panel value={ORDER.NEWEST} pt='xl'>
 				{generateProductCards()}
 			</Tabs.Panel>
+
+			<span ref={ref} />
+			{(isFetchingNextPage || isLoading) && (
+				<Loader size={'sm'} variant={'bars'} mt={'xl'} />
+			)}
 		</Tabs>
 	)
 }
